+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+- Added support for deployment strategy
+
 ## [0.0.40] - 2025-05-28
 
 ### Added